fix: usr: changed 'integer' to 'float' for sparsity type in docstrings @minor

+++ nmtf/modules/nmtf.py
@@ -463,7 +463,7 @@ class NTF:
             Select whether the regularization affects the components (H), the
             transformation (W) or none of them.
 
-        sparsity : integer, default: 0
+        sparsity : float, default: 0
             Sparsity target with 0 <= sparsity <= 1 representing the mean % rows per column in W or H set to 0
 .
         W : array-like, shape (n_samples, n_components)
